[WIP] Refactor and simplify the core logic.
Instead of working with a set of may-alias locks and unlocks, work by traversing the dominance tree in postorder (inside out).
Then match the lock with lock with the nearest post-dominating unlock (handling the defer case if it appears before).
The post-dominance and post-order of dominance code is added to the SSA package (not part of this diff).
Reduce a lot of repeated code and simplify the logic.
luPoint and luPaie are the core data structures that will be passed around.
Spefically, luPoint will be the one to be communicated between SSA and AST formats.
An luPoint will carry all info (pointer vs. value, lock vs. unlock, rlock vs. wlock, lambda vs. non lambda, etc).
Add recursive callgraph traversal, which was previously missing.
This diff disables rewriteAST, since that needs to be rewritten to work with the new logic.