Handle indirect function calls to Lock/Unlock operations.
If a lock/unlock is indirectly invoked, we do not transform them.
However, we must make sure that they work cleanly with the rest of our workflow, i.e.,
they contribute to the aliasing analysis for intra- and inter-procedural analysis.
In particular, use the indirect call to find all potential functions they may invoke and if they are
Lock/Unlock on a Mutex/RWMutex type, include all the points to set of those mutexes in the alias analysis.